requirements.txt
aiofiles==24.1.0
aiohttp==3.12.13
beautifulsoup4==4.13.4
freezegun==1.5.2
py-cord==2.6.1
pytest
pytest-asyncio==1.0.0
pytest-cov==2.6.1
python-dotenv==1.0.0
PyYAML==6.0.2
